requirements/base.txt
# Base
pytz
python-slugify
psycopg2

# Django
Django==3.2
django-cors-headers==3.7.0
django-rest-enumfield==0.2.0
django-enum-choices==2.1.3
djangorestframework-simplejwt


# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.12.4
django-filter==2.4.0

#Cognito JWT
pyjwt==1.7.1

# Passwords security
argon2-cffi==20.1.0

# Static files
whitenoise==5.2.0

# Image
Pillow==8.2.0